script folders reordering
