+++ scripts/governor/propose/_old/governance-config/index.sh
@@ -33,7 +33,7 @@ print_error() {
 show_help() {
     echo -e "${BLUE}🔧 Governance Configuration Change Script Wrapper${NC}"
     echo ""
-    echo "Usage: ./scripts/governance-config/index.sh [options]"
+    echo "Usage: ./scripts/governor/propose/_old/governance-config/index.sh [options]"
     echo ""
     echo "Options:"
     echo "  -n, --network <network>     Network to use (default: local)"
@@ -42,10 +42,10 @@ show_help() {
     echo ""
     echo "Examples:"
     echo "  # Change governance configuration on local network (interactive)"
-    echo "  ./scripts/governance-config/index.sh -n local -d dai"
+    echo "  ./scripts/governor/propose/_old/governance-config/index.sh -n local -d dai"
     echo ""
     echo "  # Change governance configuration on polygon network (interactive)"
-    echo "  ./scripts/governance-config/index.sh -n polygon -d usdc"
+    echo "  ./scripts/governor/propose/_old/governance-config/index.sh -n polygon -d usdc"
     echo ""
     echo "Interactive prompts:"
     echo "  - Admin addresses: Enter comma-separated list of admin addresses"
