+++ scripts/governor/propose/_old/timelock-delay-change/index.sh
@@ -33,7 +33,7 @@ print_error() {
 show_help() {
     echo -e "${BLUE}⏰ Timelock Delay Change Script Wrapper${NC}"
     echo ""
-    echo "Usage: ./scripts/timelock-delay-change/index.sh [options]"
+    echo "Usage: ./scripts/governor/propose/_old/timelock-delay-change/index.sh [options]"
     echo ""
     echo "Options:"
     echo "  -n, --network <network>     Network to use (default: local)"
@@ -41,10 +41,10 @@ show_help() {
     echo ""
     echo "Examples:"
     echo "  # Change timelock delay on local network (delay will be asked interactively)"
-    echo "  ./scripts/timelock-delay-change/index.sh -n local"
+    echo "  ./scripts/governor/propose/_old/timelock-delay-change/index.sh -n local"
     echo ""
     echo "  # Change timelock delay on mainnet (delay will be asked interactively)"
-    echo "  ./scripts/timelock-delay-change/index.sh -n mainnet"
+    echo "  ./scripts/governor/propose/_old/timelock-delay-change/index.sh -n mainnet"
     echo ""
     echo "Delay examples (when prompted):"
     echo "  86400  = 1 day"
